requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
langchain-openai==0.0.2
python-dotenv==1.0.0
asyncio
